Remove challenge code from OSCObjectSpace, add option to reject events
It's better to have the authentication code separate and modular, so the assumption now is that the dataspace is using an authenticated addrBook if that is needed. Also we add the option (on by default) to reject Events, as these in the form of pseudo-objects are the most likely source of malicious code.

Signed-off-by: Scott Wilson <i@scottwilson.ca>